server/requirements.txt
# Core web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# AI and Language Models
openai>=1.3.7
langchain==0.2.5
langchain-openai==0.1.14
langchain-community==0.2.5
langchain-core==0.2.8

# Document processing and chunking
# Note: Using alternative document processing libraries
pypdf==3.17.4
python-docx==1.1.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Database and Vector Search
psycopg[binary]
chromadb==0.4.24
tiktoken==0.7.0
# pgvector==0.2.4  # Will install after PostgreSQL setup
sqlalchemy==2.0.23

# HTTP requests and API communication
httpx==0.25.2
requests==2.31.0

# Environment variables
python-dotenv==1.0.0

# Data processing (compatible with Python 3.13)
pandas>=2.2.0
numpy<2.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1

# CORS for frontend integration
python-multipart==0.0.6

# JSON processing (using compatible version)
pydantic<2,>=1

# Logging
loguru==0.7.2
